fixed metric for saving. made script to update the company name file. general cleaning

+++ clean/download_company_names.py
@@ -1,11 +1,15 @@
 from selenium.webdriver.support.ui import Select
+from const import CONFIG, DIR, logger
 from selenium import webdriver
 import chromedriver_binary
 from pathlib import Path
-from const import DIR
 import pandas as pd
+import traceback
+import sys, os
 import time
-import os
+
+sys.path.append(f"{DIR}/..")
+from utils import send_metric
 
 ###################################################################################################
 
@@ -14,16 +18,19 @@ EXCHANGES = [
 	"LSE",
 	"NASDAQ",
 	"NYSE",
-	"TSX"
+	"TSX",
+	"TSXV"
 ]
 
 DATA = Path(f"{DIR}/name_data")
 
 ###################################################################################################
 
-def download_and_merge_names():
+def download_company_names():
 
 	for file in DATA.iterdir():
+		if file.name == '.gitignore':
+			continue
 		file.unlink()
 
 	###############################################################################################
@@ -39,7 +46,7 @@ def download_and_merge_names():
 
 	driver = webdriver.Chrome(options = options)
 
-	print("Getting web page...")
+	logger.info("Getting web page...")
 	driver.get("http://eoddata.com")
 
 	username_input = driver.find_element_by_id("ctl00_cph1_lg1_txtEmail")
@@ -49,15 +56,15 @@ def download_and_merge_names():
 	username_input.send_keys("zQuantz")
 	password_input.send_keys("Street101!")
 	
-	print("Logging in...")
+	logger.info("Logging in...")
 	login_button.click()
 
-	print("Getting download page...")
+	logger.info("Getting download page...")
 	driver.get("http://eoddata.com/symbols.aspx")
 
 	for exchange in EXCHANGES:
 
-		print("Downloading", exchange)
+		logger.info(f"Downloading: {exchange}")
 
 		exchange_selector = Select(driver.find_element_by_id("ctl00_cph1_cboExchange"))
 		exchange_selector.select_by_value(exchange)
@@ -74,14 +81,61 @@ def download_and_merge_names():
 
 	df = []
 	for file in DATA.iterdir():
+		if file.name == '.gitignore':
+			continue
 		df.append(pd.read_csv(file, delimiter="\t"))
 		df[-1]['exchange'] = file.name[:-4]
 
 	df = pd.concat(df)
 	df.columns = ['ticker', 'name', 'exchange']
-	return df.sort_values('ticker').reset_index(drop=True)
+	df.sort_values('ticker').reset_index(drop=True)
+
+	###############################################################################################
+
+	df = df[df.ticker.str.len() <= 6]
+
+	combo = df.name + " " + df.exchange
+	vcs = combo.value_counts()    
+	df = df[combo.isin(vcs[vcs == 1].index)]
+	df = df[df.ticker.str.count("\\.") <= 1]
+
+	ndaq = df[df.exchange == 'NASDAQ']
+	ndaq = ndaq[ndaq.ticker.str.len() > 4]
+	mods = ndaq.ticker.str[-1]
+	mods = ndaq[~mods.isin(['A', 'B', 'C'])]
+	df = df[~df.index.isin(mods.index)]
+
+	ticker_mods = df[df.ticker.str.count("\\.") == 1]
+	mod = ticker_mods.ticker.str.split("\\.").str[-1]
+	ticker_mods = ticker_mods[~mod.isin(["A", "B", "C"])]
+	df = df[~df.index.isin(ticker_mods.index)]
+
+	df = df[~df.ticker.str[0].str.isnumeric()]
+	df = df[~df.ticker.str[-1].str.isnumeric()]
+
+	mods = df[df.ticker.str.count("-") == 1].ticker
+	mods = mods[~mods.str[-1].isin(['A', 'B', 'C'])]
+	df = df[~df.index.isin(mods.index)]
+
+	return df
 
 if __name__ == '__main__':
 
-	df = download_and_merge_names()
-	df.to_csv(f"{DIR}/data/company_names.csv", index=False)
+	logger.info("company name downloader initialized")
+	metric = "company_name_download_success_indicator"
+
+	try:
+
+		df = download_company_names()
+		df.to_csv(f"{DIR}/data/company_names.csv", index=False)
+
+		logger.info("company name downloader succesful")
+		send_metric(CONFIG, metric, "int64_value", 1)
+
+	except Exception as e:
+
+		exc = traceback.format_exc()
+		logger.warning(f"company name downloader failed, {e}, {exc}")
+		send_metric(CONFIG, metric, "int64_value", 0)
+
+	logger.info("company name downloader terminated")
